Base class for DSLRunner and Optimizer
