+++ docetl/runner.py
@@ -9,10 +9,10 @@ from rich.console import Console
 from docetl.dataset import Dataset, create_parsing_tool_map
 from docetl.operations import get_operation
 from docetl.operations.utils import flush_cache
-from docetl.utils import load_config
 import pyrate_limiter
 import math
 from inspect import isawaitable
+from .pipeline import Pipeline
 
 load_dotenv()
 
@@ -37,7 +37,7 @@ class BucketCollection(pyrate_limiter.BucketFactory):
             return self.buckets["unknown"]
         return self.buckets[item.name]
     
-class DSLRunner:
+class DSLRunner(Pipeline):
     """
     A class for executing Domain-Specific Language (DSL) configurations.
 
@@ -60,7 +60,7 @@ class DSLRunner:
         Args:
             max_threads (int, optional): Maximum number of threads to use. Defaults to None.
         """
-        self.config = config
+        Pipeline.__init__(self, config)
         self.default_model = self.config.get("default_model", "gpt-4o-mini")
         self.max_threads = max_threads or (os.cpu_count() or 1) * 4
         self.console = Console()
@@ -104,11 +104,6 @@ class DSLRunner:
         bucket_factory = BucketCollection(**buckets)
         self.rate_limiter = pyrate_limiter.Limiter(bucket_factory, max_delay=math.inf)
         
-    @classmethod
-    def from_yaml(cls, yaml_file: str, **kwargs):
-        config = load_config(yaml_file)
-        return cls(config, **kwargs)
-
     def syntax_check(self):
         """
         Perform a syntax check on all operations defined in the configuration.
